mongodb-qe-tutorial/requirements.txt
pymongo>=4.4.0 
pymongo[encryption] 
python-dotenv